backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0"
description = "Backend for gshare project"
authors = ["Alexey Kutsenko <kutsenko_alexey@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.2"
PyDrive = "^1.3.1"
python-dotenv = "^0.21.0"
flask-cors = "^3.0.10"
google-api-python-client = "^2.66.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.7.1"
pymongo = "^4.3.3"

[tool.poetry.group.dev.dependencies]
Sphinx = "^5.3.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
